+++ hapi-fhir-cli/hapi-fhir-cli-jpaserver/src/main/java/ca/uhn/fhir/jpa/demo/JpaServerDemo.java
@@ -28,7 +28,7 @@ import ca.uhn.fhir.interceptor.api.IInterceptorBroadcaster;
 import ca.uhn.fhir.jpa.api.config.DaoConfig;
 import ca.uhn.fhir.jpa.api.dao.DaoRegistry;
 import ca.uhn.fhir.jpa.api.dao.IFhirSystemDao;
-import ca.uhn.fhir.jpa.binstore.BinaryAccessProvider;
+import ca.uhn.fhir.jpa.binary.BinaryAccessProvider;
 import ca.uhn.fhir.jpa.config.JpaConfig;
 import ca.uhn.fhir.jpa.interceptor.CascadingDeleteInterceptor;
 import ca.uhn.fhir.jpa.provider.JpaCapabilityStatementProvider;
@@ -185,7 +185,6 @@ public class JpaServerDemo extends RestfulServer {
 		getInterceptorService().registerInterceptor(new ResponseHighlighterInterceptor());
 
 		registerProvider(myAppCtx.getBean(BinaryAccessProvider.class));
-
 	}
 
 }
