+++ hapi-fhir-storage/src/main/java/ca/uhn/fhir/jpa/binary/BaseBinaryStorageSvcImpl.java
@@ -1,4 +1,4 @@
-package ca.uhn.fhir.jpa.binstore;
+package ca.uhn.fhir.jpa.binary;
 
 /*-
  * #%L
@@ -22,6 +22,7 @@ package ca.uhn.fhir.jpa.binstore;
 
 import ca.uhn.fhir.i18n.Msg;
 import ca.uhn.fhir.context.FhirContext;
+import ca.uhn.fhir.jpa.binary.IBinaryStorageSvc;
 import ca.uhn.fhir.rest.server.exceptions.InternalErrorException;
 import ca.uhn.fhir.rest.server.exceptions.PayloadTooLargeException;
 import ca.uhn.fhir.util.BinaryUtil;
@@ -47,7 +48,7 @@ import java.util.Optional;
 import static org.apache.commons.lang3.StringUtils.isBlank;
 import static org.apache.commons.lang3.StringUtils.isNotBlank;
 
-abstract class BaseBinaryStorageSvcImpl implements IBinaryStorageSvc {
+public abstract class BaseBinaryStorageSvcImpl implements IBinaryStorageSvc {
 	private final SecureRandom myRandom;
 	private final String CHARS = "abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789";
 	private final int ID_LENGTH = 100;
@@ -56,7 +57,7 @@ abstract class BaseBinaryStorageSvcImpl implements IBinaryStorageSvc {
 	@Autowired
 	private FhirContext myFhirContext;
 
-	BaseBinaryStorageSvcImpl() {
+	public BaseBinaryStorageSvcImpl() {
 		myRandom = new SecureRandom();
 	}
 
@@ -98,13 +99,13 @@ abstract class BaseBinaryStorageSvcImpl implements IBinaryStorageSvc {
 
 	@SuppressWarnings("UnstableApiUsage")
 	@Nonnull
-	HashingInputStream createHashingInputStream(InputStream theInputStream) {
+	protected HashingInputStream createHashingInputStream(InputStream theInputStream) {
 		HashFunction hash = Hashing.sha256();
 		return new HashingInputStream(hash, theInputStream);
 	}
 
 	@Nonnull
-	CountingInputStream createCountingInputStream(InputStream theInputStream) {
+	protected CountingInputStream createCountingInputStream(InputStream theInputStream) {
 		InputStream is = ByteStreams.limit(theInputStream, getMaximumBinarySize() + 1L);
 		return new CountingInputStream(is) {
 			@Override
@@ -118,7 +119,7 @@ abstract class BaseBinaryStorageSvcImpl implements IBinaryStorageSvc {
 		};
 	}
 
-	String provideIdForNewBlob(String theBlobIdOrNull) {
+	protected String provideIdForNewBlob(String theBlobIdOrNull) {
 		String id = theBlobIdOrNull;
 		if (isBlank(theBlobIdOrNull)) {
 			id = newBlobId();
